Fix/chunk nan overflow safe exp (#131)
* fix: prevent exp overflow in chunk_o.py for large chunk_size

When chunk_size > 64, exp(g_diff) in the masked-out region of the
attention matrix can overflow f32 (e.g. |gamma|*128 = 90.5 > 88.72).
On TPU, XLA may compile jnp.where(mask, val, 0) as mask*val, causing
0*Inf=NaN to propagate.

Fix: use jnp.where(mask, g_diff, 0.0) before exp() so masked entries
get exp(0)=1 instead of Inf. This is sign-agnostic (works for any
g_gamma, not just negative) and matches Triton's tl.where semantics.

Affected paths:
- _chunk_fwd_o_kernel (g_gamma branch)
- chunk_simple_gla_bwd_kernel (A recomputation + dA gating)
- chunk_fwd_o_ref, chunk_bwd_dv, chunk_bwd_dqkwg (reference impls)

Tested with actual ALiBi slopes (H=16, 20 layers) at chunk_size=128.


* align pallas/cpu vs tirton gpu

* fix ci error.

* Perf/gla bwd (#130)

* rm gk, rm recompute of h

* add disable_bounds_checks

* revert store h

* docs: add kernel correctness methodology with GPU/TPU error bound constraint (#134)

- Define high-precision CPU JAX (float64) as the sole Golden baseline
- GPU implementation compared against high-precision CPU JAX (ε_GPU)
- TPU implementation compared against high-precision CPU JAX (ε_TPU)
- Enforce error bound constraint: ε_TPU ≤ ε_GPU
- Add GPU/TPU test file naming convention (_gpu/_tpu suffix)

* fix ci error.

* fix ci

* fix ci

---------

+++ tops/ops/common/chunk_h.py
@@ -334,8 +334,8 @@ def _chunk_fwd_h_scan(k, v, g, g_gamma, gk, h0,
 
         # State update: h += k^T @ v  (contract C, batch B and H)
         kv = lax.dot_general(
-            b_k.astype(jnp.float32),
-            b_v.astype(jnp.float32),
+            b_k,
+            b_v,
             dimension_numbers=(((1,), (1,)), ((0, 2), (0, 2))),
             precision=lax.Precision.HIGHEST,
             preferred_element_type=jnp.float32,
@@ -547,8 +547,8 @@ def _chunk_bwd_dh_scan(q, do, g, g_gamma, gk, dht,
 
         # Accumulate: dh += q_hat^T @ do  (contract C, batch B and H)
         dh = dh + lax.dot_general(
-            b_q_hat.astype(jnp.float32),
-            b_do.astype(jnp.float32),
+            b_q_hat,
+            b_do,
             dimension_numbers=(((1,), (1,)), ((0, 2), (0, 2))),
             precision=lax.Precision.HIGHEST,
             preferred_element_type=jnp.float32,
